requirements.txt
asgiref==3.4.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2.13
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
idna==3.3
Pillow==8.4.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
